# kv: move splitHealthy to method on grpcTransport